+++ CHANGELOG.md
@@ -27,6 +27,10 @@ These are the section headers that we use:
 
 - Updated `Dockerfile` to use multi stage build ([#3221](https://github.com/argilla-io/argilla/pull/3221) and [#3793](https://github.com/argilla-io/argilla/pull/3793)).
 
+### Fixed
+
+- Fixed allow pull datasets without records ([#3851](https://github.com/argilla-io/argilla/pull/3851))
+
 ## [1.16.0](https://github.com/argilla-io/argilla/compare/v1.15.1...v1.16.0)
 
 ### Added